# marketstack-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Marketstack OpenAPI",
    "description": "Inofficial marketstack OpenAPI spec",
    "version": "1.0"
  },
  "paths": {
...
    "/intraday": {
      "get": {
        "tags": [
          "intraday"
        ],
        "summary": "Query",
        "operationId": "intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/intraday/latest": {
      "get": {
        "tags": [
          "intraday"
        ],
        "summary": "Latest",
        "operationId": "intraday_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/intraday/{date}": {
      "get": {
        "tags": [
          "intraday"
        ],
        "summary": "Date",
        "operationId": "intraday_date",
        "parameters": [
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tickers/{symbol}/intraday": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Intraday",
        "operationId": "ticker_symbol_intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_TickerIntraday_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/intraday/latest": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Intraday Latest",
        "operationId": "ticker_symbol_intraday_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntervalPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/intraday/{date}": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Intraday Date",
        "operationId": "ticker_symbol_intraday_date",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/exchanges/{mic}/intraday": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Intraday",
        "operationId": "exchange_mic_intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeIntraday_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/intraday/latest": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Intraday Latest",
        "operationId": "exchange_mic_intraday_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeIntraday_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "title": "Error",
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "context": {
            "title": "Context",
            "type": "object"
          }
        }
      },
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Interval": {
        "title": "Interval",
        "enum": [
          "1min",
          "5min",
          "10min",
          "30min",
          "1hour",
          "3hour",
          "6hour",
          "12hour",
          "24hour"
        ],
        "description": "An enumeration."
      },
...
      "PagedResponse_ExchangeIntraday_": {
        "title": "PagedResponse[ExchangeIntraday]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/ExchangeIntraday"
          }
        }
      },
...
      "PagedResponse_List_models.IntervalPrice__": {
        "title": "PagedResponse[List[models.IntervalPrice]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntervalPrice"
            }
          }
        }
      },
...
      "PagedResponse_TickerIntraday_": {
        "title": "PagedResponse[TickerIntraday]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/TickerIntraday"
          }
        }
      },
...
      "Sort": {
        "title": "Sort",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "An enumeration."
      },
...
    }
  }
}